GG EZ

+++ src/tugasmvc/Controller.java
@@ -17,7 +17,9 @@ import javax.swing.JTable;
  * @author Acer
  */
 //Pandu Dhaulagiri - 124190044 - PBO A
+
 public class Controller {
+	int mastervariable=0;
     Model model;
     View view;
     
@@ -66,17 +68,9 @@ public class Controller {
         view.jbreset.addActionListener(new ActionListener() {//Tombol Reset
             @Override
             public void actionPerformed(ActionEvent ae) {
-                String nis = view.getID();
-                String nama = view.getNama();
-                String alamat = view.getAlamat();
-                
-                if (nis.equals("") && nama.equals("") && alamat.equals("")) {   
-                } else {
-                    view.reset();
-                }
-                
-                String dataSiswa[][] = model.read();
-                view.tabel.setModel(new JTable(dataSiswa, view.namaKolom).getModel());
+                model.reset();
+                String data[][] = model.read();
+                view.tabel.setModel(new JTable(data, view.namaKolom).getModel());
             }
         });
 
@@ -84,21 +78,24 @@ public class Controller {
             public void mouseClicked(MouseEvent e) {
                 super.mousePressed(e);
                 int baris = view.tabel.getSelectedRow();
-
+                
                 int dataterpilih = (int) view.tabel.getValueAt(baris, 0);
-
+                
                 System.out.println(dataterpilih);
-
+                
                 view.jbdelete.addActionListener(new ActionListener() {
 	                public void actionPerformed(ActionEvent ae){
-	                    int input = JOptionPane.showConfirmDialog(null,"Apa anda ingin menghapus " + dataterpilih + "?", "Perhatian", JOptionPane.YES_NO_OPTION);
-	                    if (input == 0) {
-		                    model.delete(dataterpilih);
-		                    String dataMahasiswa[][] = model.read();
-		                    view.tabel.setModel(new JTable(dataMahasiswa, view.namaKolom).getModel());
-	                    } else {
-	                    JOptionPane.showMessageDialog(null, "Cancelled");
-	                    }
+	                	if(mastervariable>0) {
+	                		int input = JOptionPane.showConfirmDialog(null,"Apa anda ingin menghapus " + dataterpilih + "?", "Perhatian", JOptionPane.YES_NO_OPTION);
+		                    if (input == 0) {
+			                    model.delete(dataterpilih);
+			                    String dataMahasiswa[][] = model.read();
+			                    view.tabel.setModel(new JTable(dataMahasiswa, view.namaKolom).getModel());
+			                    mastervariable+=1;
+		                    } else {
+		                    JOptionPane.showMessageDialog(null, "Cancelled");
+		                    }
+	                	}
 	                }
                 });
             }
